# MVC/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Gestión de Tareas",
    "description": "API para gestionar tareas con autenticación JWT de usuarios",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/users/profile": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
...
}